requirements.txt
pandas
scikit-learn
mlflow
dvc
fastapi
uvicorn
pyyaml
numpy
